Use clang-format; flesh out layout.c
Started using clang-format (LLVM style) on the codebase. Implemented rudimentary versions of the functions in src/core/layout.c--still need to handle errors properly (instead of exiting early), maximize lexicographical ordering of the layouts, and implement condition (d) of the WROM algorithm. (On that note--also changed the README to use the name 'WROM algorithm'.)